refactor(tui): extract all 12 picker dispatch blocks into usePickerKeys hook (Issue #23)
